refactor(types): type generation
