+++ packages/oruga-next/package.json
@@ -49,8 +49,8 @@
     "test:ts": "vue-tsc --noEmit --skipLibCheck",
     "test:watch": "rm -rf .nyc_output coverage && NODE_ENV=test cypress open --component",
     "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --quiet --ignore-path .gitignore",
-    "gen-volar-dts": "node ../../.scripts/gen-component-declaration.mjs --bundle --platform=node",
-    "gen-comp-types": "node ../../.scripts/gen-component-types.mjs --bundle --platform=node",
+    "gen-volar-dts": "node ../../.scripts/gen-volar-dts.mjs --bundle --platform=node",
+    "gen-comp-types": "node ../../.scripts/gen-comp-types.mjs --bundle --platform=node",
     "update": "ncu -u"
   },
   "keywords": [
