+++ packages/oruga-next/volar.d.ts
@@ -1,42 +1,42 @@
 // Auto generated component declarations
 declare module "vue" {
-  export interface GlobalComponents {
-    OAutocomplete: (typeof import("@oruga-ui/oruga-next"))["OAutocomplete"];
-    OButton: (typeof import("@oruga-ui/oruga-next"))["OButton"];
-    OCarousel: (typeof import("@oruga-ui/oruga-next"))["OCarousel"];
-    OCarouselItem: (typeof import("@oruga-ui/oruga-next"))["OCarouselItem"];
-    OCheckbox: (typeof import("@oruga-ui/oruga-next"))["OCheckbox"];
-    OCollapse: (typeof import("@oruga-ui/oruga-next"))["OCollapse"];
-    ODatepicker: (typeof import("@oruga-ui/oruga-next"))["ODatepicker"];
-    ODatetimepicker: (typeof import("@oruga-ui/oruga-next"))["ODatetimepicker"];
-    ODropdown: (typeof import("@oruga-ui/oruga-next"))["ODropdown"];
-    ODropdownItem: (typeof import("@oruga-ui/oruga-next"))["ODropdownItem"];
-    OField: (typeof import("@oruga-ui/oruga-next"))["OField"];
-    OIcon: (typeof import("@oruga-ui/oruga-next"))["OIcon"];
-    OInput: (typeof import("@oruga-ui/oruga-next"))["OInput"];
-    OLoading: (typeof import("@oruga-ui/oruga-next"))["OLoading"];
-    OMenu: (typeof import("@oruga-ui/oruga-next"))["OMenu"];
-    OMenuItem: (typeof import("@oruga-ui/oruga-next"))["OMenuItem"];
-    OModal: (typeof import("@oruga-ui/oruga-next"))["OModal"];
-    ONotification: (typeof import("@oruga-ui/oruga-next"))["ONotification"];
-    OPagination: (typeof import("@oruga-ui/oruga-next"))["OPagination"];
-    ORadio: (typeof import("@oruga-ui/oruga-next"))["ORadio"];
-    OSelect: (typeof import("@oruga-ui/oruga-next"))["OSelect"];
-    OSidebar: (typeof import("@oruga-ui/oruga-next"))["OSidebar"];
-    OSkeleton: (typeof import("@oruga-ui/oruga-next"))["OSkeleton"];
-    OSlider: (typeof import("@oruga-ui/oruga-next"))["OSlider"];
-    OSliderTick: (typeof import("@oruga-ui/oruga-next"))["OSliderTick"];
-    OStepItem: (typeof import("@oruga-ui/oruga-next"))["OStepItem"];
-    OSteps: (typeof import("@oruga-ui/oruga-next"))["OSteps"];
-    OSwitch: (typeof import("@oruga-ui/oruga-next"))["OSwitch"];
-    OTable: (typeof import("@oruga-ui/oruga-next"))["OTable"];
-    OTableColumn: (typeof import("@oruga-ui/oruga-next"))["OTableColumn"];
-    OTabItem: (typeof import("@oruga-ui/oruga-next"))["OTabItem"];
-    OTabs: (typeof import("@oruga-ui/oruga-next"))["OTabs"];
-    OTaginput: (typeof import("@oruga-ui/oruga-next"))["OTaginput"];
-    OTimepicker: (typeof import("@oruga-ui/oruga-next"))["OTimepicker"];
-    OTooltip: (typeof import("@oruga-ui/oruga-next"))["OTooltip"];
-    OUpload: (typeof import("@oruga-ui/oruga-next"))["OUpload"];
-  }
+    export interface GlobalComponents {
+        OAutocomplete: (typeof import("@oruga-ui/oruga-next"))["OAutocomplete"];
+        OButton: (typeof import("@oruga-ui/oruga-next"))["OButton"];
+        OCarousel: (typeof import("@oruga-ui/oruga-next"))["OCarousel"];
+        OCarouselItem: (typeof import("@oruga-ui/oruga-next"))["OCarouselItem"];
+        OCheckbox: (typeof import("@oruga-ui/oruga-next"))["OCheckbox"];
+        OCollapse: (typeof import("@oruga-ui/oruga-next"))["OCollapse"];
+        ODatepicker: (typeof import("@oruga-ui/oruga-next"))["ODatepicker"];
+        ODatetimepicker: (typeof import("@oruga-ui/oruga-next"))["ODatetimepicker"];
+        ODropdown: (typeof import("@oruga-ui/oruga-next"))["ODropdown"];
+        ODropdownItem: (typeof import("@oruga-ui/oruga-next"))["ODropdownItem"];
+        OField: (typeof import("@oruga-ui/oruga-next"))["OField"];
+        OIcon: (typeof import("@oruga-ui/oruga-next"))["OIcon"];
+        OInput: (typeof import("@oruga-ui/oruga-next"))["OInput"];
+        OLoading: (typeof import("@oruga-ui/oruga-next"))["OLoading"];
+        OMenu: (typeof import("@oruga-ui/oruga-next"))["OMenu"];
+        OMenuItem: (typeof import("@oruga-ui/oruga-next"))["OMenuItem"];
+        OModal: (typeof import("@oruga-ui/oruga-next"))["OModal"];
+        ONotification: (typeof import("@oruga-ui/oruga-next"))["ONotification"];
+        OPagination: (typeof import("@oruga-ui/oruga-next"))["OPagination"];
+        ORadio: (typeof import("@oruga-ui/oruga-next"))["ORadio"];
+        OSelect: (typeof import("@oruga-ui/oruga-next"))["OSelect"];
+        OSidebar: (typeof import("@oruga-ui/oruga-next"))["OSidebar"];
+        OSkeleton: (typeof import("@oruga-ui/oruga-next"))["OSkeleton"];
+        OSlider: (typeof import("@oruga-ui/oruga-next"))["OSlider"];
+        OSliderTick: (typeof import("@oruga-ui/oruga-next"))["OSliderTick"];
+        OStepItem: (typeof import("@oruga-ui/oruga-next"))["OStepItem"];
+        OSteps: (typeof import("@oruga-ui/oruga-next"))["OSteps"];
+        OSwitch: (typeof import("@oruga-ui/oruga-next"))["OSwitch"];
+        OTable: (typeof import("@oruga-ui/oruga-next"))["OTable"];
+        OTableColumn: (typeof import("@oruga-ui/oruga-next"))["OTableColumn"];
+        OTabItem: (typeof import("@oruga-ui/oruga-next"))["OTabItem"];
+        OTabs: (typeof import("@oruga-ui/oruga-next"))["OTabs"];
+        OTaginput: (typeof import("@oruga-ui/oruga-next"))["OTaginput"];
+        OTimepicker: (typeof import("@oruga-ui/oruga-next"))["OTimepicker"];
+        OTooltip: (typeof import("@oruga-ui/oruga-next"))["OTooltip"];
+        OUpload: (typeof import("@oruga-ui/oruga-next"))["OUpload"];
+    }
 }
 export {};
